Hwf master v2 fix code smell (#2064)
* fix code smells

* fix code smells

* fix code smells

* fix code smells

* fix code smells

* fix code smells

* Updated HWF Fee Remission Callback Handler
