+++ src/main/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/HWFFeeRemittedCallbackHandler.java
@@ -8,7 +8,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.cmc.ccd.domain.CaseEvent;
 import uk.gov.hmcts.cmc.ccd.mapper.CaseMapper;
+import uk.gov.hmcts.cmc.claimstore.events.EventProducer;
+import uk.gov.hmcts.cmc.claimstore.idam.models.User;
 import uk.gov.hmcts.cmc.claimstore.services.DirectionsQuestionnaireDeadlineCalculator;
+import uk.gov.hmcts.cmc.claimstore.services.UserService;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.Role;
 import uk.gov.hmcts.cmc.claimstore.utils.CaseDetailsConverter;
 import uk.gov.hmcts.cmc.domain.models.Claim;
@@ -17,6 +20,7 @@ import uk.gov.hmcts.cmc.domain.utils.MonetaryConversions;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
+import uk.gov.hmcts.reform.ccd.client.model.SubmittedCallbackResponse;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
@@ -41,19 +45,28 @@ public class HWFFeeRemittedCallbackHandler extends CallbackHandler {
 
     private final CaseMapper caseMapper;
 
+    private final EventProducer eventProducer;
+
+    private final UserService userService;
+
     @Autowired
     public HWFFeeRemittedCallbackHandler(CaseDetailsConverter caseDetailsConverter,
                                          DirectionsQuestionnaireDeadlineCalculator deadlineCalculator,
-                                         CaseMapper caseMapper) {
+                                         CaseMapper caseMapper,
+                                         EventProducer eventProducer,
+                                         UserService userService) {
         this.caseDetailsConverter = caseDetailsConverter;
         this.deadlineCalculator = deadlineCalculator;
         this.caseMapper = caseMapper;
+        this.eventProducer = eventProducer;
+        this.userService = userService;
     }
 
     @Override
     protected Map<CallbackType, Callback> callbacks() {
         return ImmutableMap.of(
-            CallbackType.ABOUT_TO_SUBMIT, this::updateFeeRemitted
+            CallbackType.ABOUT_TO_SUBMIT, this::updateFeeRemitted,
+            CallbackType.SUBMITTED, this::startHwfClaimUpdatePostOperations
         );
     }
 
@@ -88,6 +101,21 @@ public class HWFFeeRemittedCallbackHandler extends CallbackHandler {
             .builder()
             .data(dataMap)
             .build();
+    }
+
+    private CallbackResponse startHwfClaimUpdatePostOperations(CallbackParams callbackParams) {
+        Claim claim = caseDetailsConverter.extractClaim(callbackParams.getRequest().getCaseDetails());
+        logger.info("Created citizen case for callback of type {}, claim with external id {}",
+            callbackParams.getType(),
+            claim.getExternalId());
+        String authorisation = callbackParams.getParams().get(CallbackParams.Params.BEARER_TOKEN).toString();
+        User user = userService.getUser(authorisation);
+        eventProducer.createHwfClaimUpdatedEvent(
+            claim,
+            user.getUserDetails().getFullName(),
+            authorisation
+        );
+        return SubmittedCallbackResponse.builder().build();
     }
 
 }
