+++ src/main/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/HWFMiscellaneousCallbackHandler.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks;
 
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -23,6 +22,7 @@ import uk.gov.hmcts.reform.ccd.client.model.SubmittedCallbackResponse;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -36,7 +36,9 @@ public class HWFMiscellaneousCallbackHandler extends CallbackHandler {
 
     private static final List<Role> ROLES = Collections.singletonList(CASEWORKER);
 
-    private static final List<CaseEvent> EVENTS = ImmutableList.of(CaseEvent.MISC_HWF);
+    private static final List<CaseEvent> EVENTS = Arrays.asList(CaseEvent.MISC_HWF,
+        CaseEvent.MORE_INFO_REQUIRED_FOR_HWF, CaseEvent.FULL_REMISSION_HWF_REJECTED,
+        CaseEvent.PART_REMISSION_HWF_REJECTED, CaseEvent.HWF_NO_REMISSION);
 
     private final CaseDetailsConverter caseDetailsConverter;
